backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
tortoise-orm>=0.19.0
fastapi-admin>=1.0.0
aioredis>=2.0.0
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
httpx>=0.24.0

# 开发/测试依赖
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx
aerich>=0.7.0
